Fixed multiple 2D Edge Collider Issue
Problem: There were 2 Edge Collider 2D on the terrain, the second one (which is diabled now) was a trigger, but it was not set to the ground for some reason. Unticking the parameter allowed for it disappear and the "Death by Head" handler is working as well. Needs thorough testing